web/requirements.txt
bokeh==1.0.4
confusable-homoglyphs==3.2.0
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.1.5
django-registration==3.0
django-toolbelt==0.0.1
djangorestframework==3.9.1
gunicorn==19.9.0
Jinja2==2.10
libsass==0.15.1
MarkupSafe==1.1.0
nltk==3.4
numpy==1.16.0
packaging==18.0
Pillow==5.4.1
psycopg2==2.7.6.1
psycopg2-binary==2.7.6.1
pyparsing==2.3.1
python-dateutil==2.7.5
pytz==2018.9
PyYAML==3.13
singledispatch==3.4.0.3
six==1.12.0
static3==0.7.0
tornado==5.1.1
